Fix BT noise issue during startup
[Description]
For SBC codec, when PCM data isn't enough to encode,
it will be saved as residual data and fill it up next time.

However, there is a bug in the existing code that clears the buffer every time, result the residual data being cleared.

[Solution]
Only when there is no residual data can the buffer be cleared.

Bug: 335785333

Test: SBC stress-test pass

Change-Id: I67951187792eb12abf6cc8e975011cfaa2524658