Revert "gd/os/log.h: Use liblog to control logging level"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b9bc5e6-4d52-4541-b28f-0c2e78d61150). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I219549f23680d6e94770ffef2a8dc4b6d05cd89e