Revert "floss: Update syslog filtering for libbluetooth"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b9bc5e6-4d52-4541-b28f-0c2e78d61150). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iab57883a1f4e81e1db10036f216ef4da5ac9703d