leaudio: When dual mode is enabled, lets don't keep CISes on suspend
Normally Bluetooth keeps CISes for some time after stream is suspended
as there are use cases when just after Suspend the Resume arrives.
This can happen e.g. on touch sounds.

However, in the case of dual mode scenario, when same device is
switching between A2DP and LE Audio, it is important to drop CISes as
fast as possible to allow controller to switch to A2DP offloader case.

For this reason timeout Suspend timeout is dropped when dual mode is
enabled.

Bug: 325404067
Test: mmm packages/modules/Bluetooth
Test: manual verification
Test: atest bluetooth_le_audio_client_test
Flag: Exempt, regression verified with unit tests
Change-Id: I1456f5e58322418486add78ac5917a67e298b6f5