btm_sec: Don't trigger new authentication before collision timer expired
After an authentication conflict, the host starts a backoff mechanism,
allowing the peer device to continue the authentication process.
However, if the host sends a new authentication request during this
period, a new au-rand will be sent to peer which may arrives at peer
before it finishing authentication, potentially causing the peer to
cease its response.

Bug: 329221156

Test: m com.android.btservices
Test: Manual | Simultaneously initiate classic Bluetooth pairing or reconnection from both the local and remote devices.

Change-Id: Ic2604ec603edb2550297ad48ac28317cdf262d9f