Revert "bluetooth_audio_hal_version_test: Use static audio"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4ee6dd2f-cbf3-44c3-87f4-4fe654b93af8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ief69c846fa60f036a0a83f0f392b8da9c6a5aeb2