Revert "le_audio: Move ASRC implementation to new direction syst..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/08081da5-30a4-4756-9457-bb1a835fc15e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib5a7b1e87e0d010423c82a13a7fca7e9b91da1c6