Remote control cannot work after HID Agent register app due to Tv don't send acl disconnect command to indicate  bluetooth disconnection, so that RCU cannot switch into irda. Terminate the link if received BTA_HH_CLOSE_EVT.
Bug: 316056335
Test: atest testRegistration
1. Remotes & Accessorise -> Pair accessory -> Pair RCU
2. HID Agent -> Enable -> New Device
[Observed result]
The indicator light flashes green, but it does not function when any RCU button is pressed.
[Expected result]
G20 turns into irda
Similar to the Bluetooth HidDevice test case in CtsVerifier.
After clicking "Register app" in the test case, the RCU will become ineffective and it will not be possible to switch to the infrared mode.
Change-Id: I746f1392ef9095612026ad34ee7cf5cf05adb424

Change-Id: I3a7c3aa63e2f04e58f7dada6deb84e28e31e435f