Replace a2dp co-exist check with inidividual a2dp sysprop check
Problem: There is an additional sysprop to indicate if source & sink
profile can co-exist or not. The sysprop is an unnecessary property
Also, enabling the sysprop cannot be done in concurrent to the trunk
flag.

Solution: Instead use the source & sink sysprop to infer whether both
profiles can co-exist or not

Test: atest net_*
Bug: 329138112
Bug: 321806163
Change-Id: I4a79ed6cedc6a42abfe733e6d0805c8e6d73021c