Floss: remove debug logs in |sco_pkt_read_ptr|
The logic here is to drain all available data in the queue until it is
empty, and there is a debug log whenever an attempt to reading the queue
results in insufficient data. This spams report logs really bad because
at every cycle of SCO packet exchange there is this log.

This CL removes the log, with the rationale that if there is ever
insufficient data, the cause elsewhere will cover the logs.

Bug: 329298322
Test: m Bluetooth
Flag: EXEMPT floss only changes
Change-Id: I3439ca6a0ae2253c7c1b48db559b727923ace1e7