Add device/generic/goldfish/audio to visibility
android.hardware.audio@7.1-impl.rahchu which is in
device/generic/goldfish/audio requires audio.bluetooth.default.
Add the package path of the audio HAL impl to the visibility property of
bluetooth HAL here.

This is in preparation for an upcoming change in Soong which handles the
installation of required modules.

Bug: 321626681
Test: builds
Change-Id: I88044bf924f109994cf039b0f1c47fb60abaa7b5