Asrc: fix threading issue in using asrc from source HAL
When the asrc is instantiated from the worker thread in the source HAL,
its scheduler (|read_clock_timer_|) is owned by the worker thread and
invocations to |SchedulePeriodic| should be run in the same thread.

This CL modifies the signature so that it is aware of the thread in
which the task should be run from.

Bug: 331119124
Test: m Bluetooth
Flag: EXEMPT changes within a flag (run_ble_audio_ticks_in_worker_thread) for source_hal / no logic change in hearing_aids
Change-Id: Ia66c25c466af878bfd77a0085e64c20e09b9b5fe