Fix handling for coded and all supported PHY
aosp/2942289 incorrectly passed an enum rather than a bitmask for
supported phy types to the native layer, and also lacked correct
handling for ALL_SUPPORTED.

Fix these issues, and add a test.

A new flag bug should not be required; these changes are a no-op
if phy_to_native is not enabled.

Test: atest ScanManagerTest
Bug: 228268706
Bug: 326249470
Change-Id: I8a385c00b1779139e2e6ac63c77434e0c497ea25