[le audio] Avoid inactivating the device while receiving broadcast
When device updated available context types once syncing to BIS, the
assistant set device to be inactive. This caused calling handover won't
happen when the sinks are receiving broadcast.
We're skipping the inactivating action if the remote devices are
receiving broadcast.

This change should only apply for assistant standalone case, for local
broadcast, we are doing the handover to stop unicast stream first so
this logic won't be triggered.

Bug: 330691253
Bug: 329188955
Bug: 308171251
Test: atest LeAudioServiceTest BassClientServiceTest
Test: manual test with broadcast handover cases
Change-Id: I1b9921a03a667d8527306e5fe6ebdc012b13161b