Revert "HidHost: Extract BluetoothDevice from the message correctly"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/157f71ea-b774-46d8-9beb-96a3592ff93f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie5d515ab06dbaddab6cdaaf3da223674bf89e3c4