Floss: Prevent phone state change when answering/hanging up by HF.
Previously, we combined the UHID input report and phone state change logic, but this caused an issue.
When the network is unstable, the call event may not be correctly propagated to RingCentral, resulting in a desynchronization between the RingCentral call state and the headset call state.
This desynchronization lead headset cannot send terminate call event because the call is already ended at headset side.
To resolve this issue, we have modified the logic to only propagate UHID input events and wait for the application to confirm the call state change via UHID output events.

Bug: 331875859
Test: Conduct the following manual tests
  pair Bluetooth Headset
  Google Meet: headset is able hang-up the call
Test: Conduct the following manual tests
  pair Bluetooth Headset
  login Ring Central and place an incoming-call
  answer the call by headset
  hangup the call by headset
Test: Conduct the following manual tests
  pair Bluetooth Headset
  login Ring Central and place an incoming-call
  answer the call on Ring Central
  hangup the call from headset
Test: atest bluetooth_test_gd
Tag: #floss
Flag: EXEMPT, no behavior change on Android; Floss-only changes

Change-Id: I76435fa424f071f50a8dc79f4ac8446db93dedcf