Floss: Implement Ring Central reject call functionality via Drop HID event
Previously, RingCentral calls could not be rejected from
a Bluetooth headset, but could be rejected from a USB headset.

The root cause is that USB headsets send an additional Drop
HID event to RingCentral to reject the call, but Floss did not.

We used to think hook-switch=0 is good enough to represent
rejecting the call in libhidtelephony. Eventually find out it
is not working on RingCentral.

Fixes the problem by having Floss also send a Drop HID input
event to reject incoming RingCentral calls.

Bug: 329304145
Test: Conduct the following manual tests
  pair Bluetooth Headset
  login Ring Central and place an incoming-call
  reject the call by headset
Test: atest bluetooth_test_gd
Tag: #floss
Flag: EXEMPT, no behavior change on Android; Floss-only changes

Change-Id: I2307df0e22cf0e10bb295c0f812416663f553d88