vc: Remove internal operation cancel
Volume control had an internal operation cancel mechanism which was
intended to unblock the queue with outstanding VC operations.

However, it turns out it does not makes sense, because if operation will
be canceled, new operation will be put on the GATT queue with
old/possibly invalid change counter.

So we tried to recover from remote issue but it will not work.

From now on, internal timer is used just to monitor on how long it takes
to process operation and helps to debug issues.

This patch also adds some stress test added when trying to understand
this issue.

Bug: 328369455
Test: atest bluetooth_vc_test
Flag; Exempt, no functional change.

Change-Id: I8f3e24a0ac28780c3728d73fc8b854412160d402