Revert "Integrate hfp_client_interface_host.cc to floss build to..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/96c663ee-7cdb-4694-a948-67385e617ba7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa483e382ab7e67d5c00cd4038536a74c52af9ea