GATT: Add context switch during handling GATT indication/notification.
After a BLE connection is established, if the first ACL packet is a GATT
indication or notification and arrives before the remote version
information completed event, the L2CAP layer will notify the connection
up event and then send the packet to the ATT (Attribute Protocol)
channel state machine. Both the connection up event and GATT
indication/notification are handled in the main thread, but the process
differs. There's a context switch for the connection up event, while the
indication/notification are handled directly in the main thread. As a
result, the indication/notification will be processed before the
connection up event and ultimately ignored because the GATT state
machine hasn't switched to the CONNECTED state yet.
Do same context switch for indication/notification.

Test: m com.android.btservices
Test: Manual | After disconnect with GATT server, change services on
server side, then trigger reconnection.

Change-Id: I674d2fd9a68e1e9c8cc87b7201d58c5f2306788b