Revert "Floss: Implement Pandora L2CAP profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6f5ee545-4b44-400d-b8ed-1a0dcc450178). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id63647b0311376489192c816e4de6c18b9199bae