Notification: Display them twice instead of once
For a while the notification where silenced by other service. Putting
the limit to twice will allow user that never seen this notification to
see it for the first time.
Also it seems to be only fair to shows the notification a couple times
before not showing forever.

In the futur we could consider showig the notification once every 6
months or so, to be sure the behavior is known without sending too many
notification

Bug: 331896688
Flag: Exempt, simple counter check
Test: m . | Notification has never been tested
Change-Id: I0174ea4e5ee2f1eb36880bcac04c4a1f64e7b526