Randomize media ID to correct lost audio focus
When an active device and secondary device are connected,
if the secondary device is disconnected and reconnected, the
active device will lose audio focus. This is because device
media IDs are stored and function is called when the secondary
device reconnects as it is a recognized device, which causes the
active device to lose audio focus and stop playing media, if any.

Tag: #stability
Bug: 215263047
Ignore-AOSP-First: code review needed
Test: atest BrowseTreeTest.java, atest AvrcpControllerStateMachineTest.java
Change-Id: Ie5e35389f55332e65f32e52162906ccf73a2ef54