Floss: notify exact group stream status to audio server
As being pointed out in a TODO comment in `le_audio_client_test.cc` that
"We should have a way to confirm Stop() otherwise, audio framework might
have different state that it is in the le_audio code", the audio server
in CrOS needs to be notified when the group is released (i.e., CIS is
teared down) so that it is allowed to start another one.

`GroupStreamStatus` appears to be very close to meeting the demand,
despite the API collapses all states to either `STREAMING` or everything
else as `IDLE`. This masks the `RELEASING` state as `IDLE` while we need
to reliably distinguish between the two states (we cannot establish
CISes while it is in `RELEASING` state).

Since it's unclear how this might affect subscribers of this callback
in Android, while there is no other more suitable alternative for Floss,
we make this change for Floss-only.

Bug: 317682584
Test: m Bluetooth
Change-Id: Ib5b61f1dd3fe15c7b07ab81b67ea416c71ee58bc