LeAudio: Split unicast and broadcast offload capabilities
Instead of using a single list for unicast and broadcast capabilities
in the bt stack, apply the device count criteria a bit earlier on the AIDL
data types and prepare a separate list for broadcast capabilities only.
We eventually want to remove the dependency to the device count from the
configuration struct in the stack and depend on the ASE count requirements
instead (like the AIDL config provider does).

Bug: 308428860
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test  bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine bluetooth_le_audio_codec_manager_test
Flag: EXEMPT; non-functional refactor; container split verified by the unit tests
Change-Id: Id7e0ba5f39402b568798cff02689cb4e4541f195