LeAudio: Add metadata LTV support
This is the proper support for the LTV data types for the LE Audio
metadata. Until now the metadata was mostly ignored, and not
interpretted. With this change the LTV Map is extended with metadata
support just like it supported the Codec Capabilities and Codec
Configuration types.

This is needed to transparently pass metada over the AIDL interface.

Bug: 340400084
Test: atest --host bluetooth_le_audio_test
Flag: EXEMPT; Data type for the multi-codec feature which is not yet enabled
Change-Id: I8fcd4bdc47685ff9fc56637deebe507f89247146