gatt: Fix handling direct connect
Fix is for the code under the flag "gatt_reconnect_on_bt_on_fix".

With this fix, gatt_if which is initiating the connection will be
properly set as a holding application after connection is established.

This was missed in the initial implementation which cause a problem with
disconnecting ACL when flag was enabled.

Bug: 325438660
Bug: 325595120
Test: atest pts-bot:GATT
Change-Id: Ib37eee6c45a094f7ee2c53b25daf39717903e3b6