Revert "Split service search result from name read result"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/73c144cd-dcd3-4456-9075-a7800e5bf882). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I19dcde3db25d5e550d45ca88a7b127479af4541e