Revert "stack::btm:: Consolidate test fixtures btm_test_fixtures.h"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e4f6fd6-f4ff-471d-9d01-c93234ef9c9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61553e26f57294e9a3a756231e9950bce118bd41