Revert "SystemServer X FastBind: prevent double enable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3f40d51b-944b-43b2-84a3-9fae8474dcf5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I012d2fc1c6742a3afb9c235963fa035d7c751867