Revert "Fix condition in device address mapping removal"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b98a1b0-2b2a-48e2-a808-c39bcd7f14ab). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie52742eccf27fa7f58b400a27d99bf29a6404f17