Revert "SystemServer X FastBind: prevent double enable"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/968f911a-6a4a-4b5e-9c9c-405d9665bd06). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibd4dfcc4cb668beee851523957000219005a85fc