Do not remove duplicate non-LE-only device records
If identity address of the newly bonded device matches with an existing
bonded device, the existing bonded device is removed. This logic was
added to accomodate some LE-only devices which keep changing their IRK
which resulted in duplicate device records.
However this logic also removes the dual mode devices which don't
support CTKD. Pairing on both transport results in separate bond state
events which triggers this logic and removes the bond for both the
transports.

Test: mmm packages/modules/Bluetooth
Test: Manual | Pair on both transport with non-CTKD supporting dual mode
devices
Test: Manual | Repair with LE-only devices which keep changing its IRK
Bug: 322360482
Bug: 333949759

Change-Id: I48fdd9be2f76b854c47f47fb57794584468dbf16