Floss: Fix Microphone mute becomes an infinite loop when phone mute report is sent twice with a value of 1
The input value for phone mute in the telephony report
descriptor was defined as 0x23, which is incorrect.

According to HID Usage Tables 1.5 -> 3.4.2 Usage Types
(Data), the correct value should be 0x22.

0x22 defines phone mute as Data. 0x23 represents Const.
Const refers to a fixed and read-only value. Data implies
that the value can be changed and is dynamic in an input
report.

Therefore, phone mute should belong to Data since its
value can be modified.

The incorrect value caused Meet to interpret the phone
mute function incorrectly,leading to an infinite loop
when the phone mute report is sent twice with a value of 1.

Bug: 333976298
Test: Meet microphone don't enter infinite loop state when the phone mute report is sent twice with a value of 1
Test: atest bluetooth_test_gd
Tag: #floss
Flag: EXEMPT, no behavior change on Android; Floss-only changes

Change-Id: Id6c66c7d955e9adfbc024266929042ea1c4544a3