Floss: Workaround for RingCentral microphone mute toggle control
When sending a 0x09, 0x2f (Phone Mute) input report to RingCentral, if the value is 0 (meaning unmute), RingCentral will not react.

The root cause is RingCentral treats 0x09, 0x2f (Phone Mute) as a toggle event instead of OOC (on/off) control. Therefore, only sending 1 with 0x09, 0x2f (Phone Mute) can toggle the microphone status on RingCentral. 0 will be treated as no-ops.

Based on the HID spec, Phone Mute belongs to OOC. This means when the input report sends 0, it means turn off the microphone. If the input report sends 1, it means turn on the microphone.

Sending this will not break the original OOC use case because when the headset sends unmute, it means the original mute state is muted. In that case, sending an additional mute event will become no-ops for the OOC use case. However, it will fix the use case for toggle.

The workaround should be removed after RingCentral implements their mute state as OOC instead of toggle.

Bug: 333847153
Bug: 329280858
Test: Conduct the following manual tests
  pair Bluetooth Headset
  Google Meet: headset is able mute/unmute via headset and the mute status sync with Meet
  RingCentral: headset is able mute/unmute via headset and the mute status sync with RingCentral
Change-Id: I6bd95681c9f58b426aef4fe630a0952c5ffaad44