Revert "Prevent variables with type java.lang.Boolean"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89ec0007-733d-4ece-9585-1c5db21c3faf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I40fb395ad6cf49fde4ef4528d83035f65e8a42cf