LeAudio: Move the SWB support toggle handling to Codec Manager
The json config provider is just the data source, from where the
configurations are loaded at the stack initalization. SWB toggle handling
should be handled by the higher level CodecManager at run time, to filter
out the SWB configuration when support is not currently enabled.

This simplifies the SWB filtering and  hides it behind the CodecManager.
This is aligned with the expected behavior for the AIDL configuration provider.

Bug: 308428860
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test  bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine bluetooth_le_audio_codec_manager_test
Flag: EXEMPT; no functional change - moved functionality verified with the unit tests
Change-Id: I96d0deb2f910a68f55ac85c214f7e0875df3d2df