Remove device properties after BOND_NONE is processed
aosp/2974117 and aosp/3038273 introduced a race condition in processing
of the BOND_NONE state change. These changes ensured that the device
mapping and device properties were removed when the bond was removed.
However, BondStateMachine relies on the device properties to decide if
the bond state change should be broadcasted. If the device properties
are removed before the BondStateMahcine handles the BONDING_STATE_CHANGE
message for the BOND_NONE state, the state change is not broadcasted.

Test: mmm packages/modules/Bluetooth
Test: Manual | Unpair a device
Flag: com.android.bluetooth.flags.remove_bond_with_address_map
Bug: 335354378
Bug: 332626602
Bug: 326294532
Bug: 326603271
Change-Id: I1e23f5d6250d9fd478262742b87ae02ab517392b