Cache UHID_INPUT event before UHID_OPEN
After the BLE connection is encrypted, HOGP device should be able
to send key notification if it is enabled in CCCD. However after
reconnection, the UHID device is not opened to be able to handle the
notification. So we need to queue the UHID_INPUT events when UHID
device is not opened and send them all at once when UHID is opened.

Note: UHID_OPEN is not an accurate indicator since it indicates a thread
has opened the uhid device but it might not be Android Input. So the
current workaround is to wait for some time after UHID_OPEN has received.

See https://kernel.org/doc/Documentation/hid/uhid.txt

Change-Id: Ifc3253a28e65a60269f8c89db59893e7172982df