Don't replace an existing CoD with one that is unclassified
Inquiry results coming from BLE can have an inferred device class based
on the service uuids or appearence. When they can't determine the device
class through their inference rules, they return a default value of
0x001f00, or uncategorized. Without a check, we end up replacing a valid
CoD with this invalid one. This can happen very easily with dual mode
devices while we're discovering. This causes the class of device to
change in Settings from the real value, to the Setting's default of
"headset."

Bug: 311456944
Bug: 335909751
Test: m com.android.btservices
Test: Pair an iOS device. Become discoverable again briefly and check
the device details icon.

Change-Id: If5fa38103a498dff491cffd27cd3094d17f973e6