Add device_type information for remote name lookup
For reading the remote device's name we can do it the BREDR way (with
HCI command remote name request), or the LE way (via GAP/GATT).
However, we don't record which way we do it, so when we need to cancel
the request, we can only infer them via the recorded device type.
Problem arises when the peer device is a DUAL device. If we cancel the
wrong type then we might fail and get stuck.

This CL saves whether we does the reading via BREDR or LE, so we don't
need to infer the device type.

Bug: 336478597
Test: m -j
Test: With a lot of fastpair-enabled peer devices simultaneously
      advertising, repeatedly do normal pairing and forget device to
      each peer device. Verify no failures because we fail to cancel
      the remote name request with the following error log:
      "Attempted to cancel remote name request to xx:xx:xx:xx:xx:xx
      that does not exist"

Change-Id: I4809f8193e7b6a4d1b5a7c68a9fc7e6f413e1f58