Revert "hci/hal: Separate AIDL/HIDL in the HAL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47f50e3d-792a-4975-b0eb-d2b62872df2e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaacd183933ec02a027c5f0fa0bfab8c1c64383a0