Revert "bt_snoop: Add intermediate copy preventing source packet..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47f50e3d-792a-4975-b0eb-d2b62872df2e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iac19e43415cac8f536b0f736eae4eee2f4e90abb