Revert "hci/hal: Move "Android" dependancies into backend"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47f50e3d-792a-4975-b0eb-d2b62872df2e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8325b9d858e9e67dde81eeedd13bf180ccbf70d5