Revert "hci/hal: Move "Android" dependancies into backend"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2713c0aa-0764-49e5-b92f-ee8069080858). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3fef93ced5a0820f39e38885932da02119a28019