Revert "bt_snoop: Add intermediate copy preventing source packet..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2713c0aa-0764-49e5-b92f-ee8069080858). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If7f0b958a8980b713490a8e45daf5dc22ee5eee8