Revert "hci/hal: Separate AIDL/HIDL in the HAL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2713c0aa-0764-49e5-b92f-ee8069080858). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6958f60717af1bc2427cd657c2bd918299d0f0c6