Uncache the current browsed player when we change browsed players
This change fixes a bug where we don't uncache a player object when we
switch to a new player. This keeps us from ever being able to get fresh
player contents when things get updated and we haven't received an
available players changed event. This also doesn't fit in with our
current caching scheme.

Bug: 337259487
Bug: 337257397
Test: m AvrcpControllerStateMachineTests
Change-Id: I36b0120dfe418b9a220ffaed8406e8c3d0fbd0e4