Revert "LeAudio: Fix redundant audio hal stop"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/172637e0-1cb4-4613-9b47-8484dc603e1e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib8ba0eadf7ae0cd52d655604ce3b5a0944866cd4