leaudio: Implement preferred config update logic
This patch implements the preferred config update logic. When setting
preferred config, we put it into requirements and find the first match
with ase config. After that with this preferred ase config, we could use
that to match with remote pac.

Multi Codec would also benifit from pushing preferred codec into
requirements as in the aidl layer they will try to return stack with the
one matched with preferred config (requirements).

Bug: 259191520
Bug: 353909820
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_le_audio_codec_manager_test
Change-Id: I8da2938d8e40fe38d3d83f6d09326f4f8fcb5ff4