leaudio: Implement preferred config update logic
This patch implements the preferred config update logic. When setting
preferred config, it iterates all contexts and check if there is one
config that is matched with preferred config for each context. If so,
we save it into `context_to_preferred_configuration_cache_map`.

Also update preferred cached when update legacy cache.

Bug: 259191520
Bug: 353909820
Test: m
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
Change-Id: I8da2938d8e40fe38d3d83f6d09326f4f8fcb5ff4